server/requirements.txt
# ChunkSmith Hybrid - Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Validation
pydantic>=2.0.0

# PDF Extraction
pymupdf>=1.23.0

# OpenSearch
opensearch-py>=2.4.0

# OpenAI SDK (for OpenAI-compatible endpoints)
openai>=1.3.0

# LangChain OpenAI embeddings
langchain-openai>=0.1.0

# File Upload
python-multipart>=0.0.6

# .env support
python-dotenv>=1.0.0
